+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v 1.9.0 (?)
 Changes in this release:
 
+- Add support for installing v2 modules in editable mode.
+
 # v 1.8.0 (2022-07-14)
 Changes in this release: