+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v 1.9.0 (?)
 Changes in this release:
 
-- Add support for installing v2 modules in editable mode.
+- We now automatically install all V2 modules found in de library path in editable mode on the remote orchestrator
 
 # v 1.8.0 (2022-07-14)
 Changes in this release: