+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Changes in this release:
 
 - We now automatically install all V2 modules found in de library path in editable mode on the remote orchestrator
+- Fix legacy option usage for `lsm_noclean` and `lsm_ssl`.
 
 # v 1.8.0 (2022-07-14)
 Changes in this release: