+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v 1.9.0 (?)
 Changes in this release:
 
-- We now automatically install all V2 modules found in de library path on the remote orchestrator from source.
+- We now automatically install all V2 modules found in de library path in editable mode on the remote orchestrator
 - Fix legacy option usage for `lsm_noclean` and `lsm_ssl` (introduced in 1.6.0).
 
 # v 1.8.0 (2022-07-14)