+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 Changes in this release:
 
 - Added `--lsm-partial-compile` option to enable partial compiles the remote orchestrator (for supported versions)
-- We now automatically install all V2 modules found in de library path in editable mode on the remote orchestrator
+- Added support for testing v2 modules: the module being tested, as well as v2 modules in the libs dir are synced to the
+    remote orchestrator and installed in editable mode.
 - Fix legacy option usage for `lsm_noclean` and `lsm_ssl` (introduced in 1.6.0).
-- Don't delete unittest module on remote orchestrator when `--use-module-in-place` is used
+- Sync all module sources to the remote orchestrator rather than only one
 
 # v 1.8.0 (2022-07-14)
 Changes in this release: