+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v 1.9.0 (?)
 Changes in this release:
 
-- Added `--lsm-partial-compile` option to enable partial compiles the remote orchestrator (for supported versions)
+- Added `--lsm-partial-compile` option to enable partial compiles on the remote orchestrator (for supported versions)
 - Added support for testing v2 modules: the module being tested, as well as v2 modules in the libs dir are synced to the
     remote orchestrator and installed in editable mode. Dependencies are installed from package sources configured through
     the `INMANTA_MODULE_REPO` environment variable or the `--module-repo` option.